Clean: Merge STEX and STU Grid
